Create SessionPage.
